requirements.txt
Django>=5.1
python-telegram-bot>=21.0
python-dotenv>=1.0.0
